Make `Cask::Auditor` return a `Hash` with warnings and errors.
