+++ Library/Homebrew/cask/all.rb
@@ -9,7 +9,6 @@ require "cask/cache"
 require "cask/cask"
 require "cask/cask_loader"
 require "cask/caskroom"
-require "cask/checkable"
 require "cask/cmd"
 require "cask/exceptions"
 require "cask/installer"
