+++ Library/Homebrew/cask/auditor.rb
@@ -2,7 +2,6 @@
 
 module Cask
   class Auditor
-    include Checkable
     extend Predicable
 
     def self.audit(cask, audit_download: false, audit_appcast: false,
@@ -39,18 +38,27 @@ module Cask
                    :audit_strict?, :audit_new_cask?, :audit_token_conflicts?, :quarantine?
 
     def audit
+      warnings = Set.new
+      errors = Set.new
+
       if !Homebrew.args.value("language") && language_blocks
-        audit_all_languages
+        language_blocks.each_key do |l|
+          audit = audit_languages(l)
+          puts audit.summary
+          warnings += audit.warnings
+          errors += audit.errors
+        end
       else
-        audit_cask_instance(cask)
+        audit = audit_cask_instance(cask)
+        puts audit.summary
+        warnings += audit.warnings
+        errors += audit.errors
       end
+
+      { warnings: warnings, errors: errors }
     end
 
     private
-
-    def audit_all_languages
-      language_blocks.keys.all?(&method(:audit_languages))
-    end
 
     def audit_languages(languages)
       ohai "Auditing language: #{languages.map { |lang| "'#{lang}'" }.to_sentence}"
@@ -71,8 +79,7 @@ module Cask
                               quarantine:      quarantine?,
                               commit_range:    commit_range)
       audit.run!
-      puts audit.summary
-      audit.success?
+      audit
     end
 
     def language_blocks
